Add cobra to the mix
Prior to this commit the the collector had a single entrypoint, main.go.
The trade-off here was convenience over scalability. It meant that we
could not easily add new functionality to the app without changing the
collector logic.

This commit adds cobra so that we can divide functionality up in to
individual commands. The benefit of this is that it also makes expansion
easier in the future.